+++ tests/Integration/ContainerTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Tests\Integration;
+namespace Tests\Integration;
 
 use App\Application;
 use App\Controller\ReservationController;